refactor(auth): rename user provider files and update imports for user state management
